move around rabbitmq client
