+++ packages/service-library/src/servicelib/rabbitmq/__init__.py
@@ -1,0 +1,3 @@
+from typing import Final
+
+BIND_TO_ALL_TOPICS: Final[str] = "#"
