+++ packages/service-library/src/servicelib/rabbitmq/constants.py
@@ -1,0 +1,3 @@
+from typing import Final
+
+REGEX_RABBIT_QUEUE_ALLOWED_SYMBOLS: Final[str] = r"^[\w\-\.]*$"
