+++ packages/service-library/src/servicelib/rabbitmq/errors.py
